> Implemented core error handline, HTTP cookie management, request/response tracking, modularised tools
