+++ bzauth-rs/Cargo.toml
@@ -12,13 +12,15 @@ runtime_actix = []
 
 [dependencies]
 # Core dependencies
-tokio = { version = "1", features = ["full"] }
-serde = { version = "1.0.218", features = ["derive"] }
+tokio = { version = "1.45.1", features = ["full"] }
+serde = { version = "1.0.219", features = ["derive"] }
+serde_json = "1.0.140"
 oauth2 = "5.0.0"
+reqwest = { version = "0.12.15", features = ["json"] }
 
 # Supported runtimes
 # todo: actix-web ? etc.
-axum = { version = "0.8.1", features = ["macros"], optional = true }
+axum = { version = "0.8.4", features = ["macros", "json"], optional = true }
 
 # Data adapters
 # todo: sqlx ? etc.
@@ -38,5 +40,9 @@ paste = "1.0.15"
 chrono = "0.4.40"
 http = "1.3.1"
 url = "2.5.4"
+rand = "0.9.1"
+tracing = { version = "0.1.41", features = ["std"] }
+tracing-subscriber = { version = "0.3.19", features = ["std"] }
+# rand_distr = "0.5.1"
 
 [dev-dependencies]
