--- bzauth-rs/src/cogs/csrf.rs
@@ -1,5 +1,0 @@
-use super::{CoreError, request::CoreRequest, response::CoreResponse};
-
-pub async fn csrf<Payload>(_request: CoreRequest<Payload>) -> Result<CoreResponse, CoreError> {
-    todo!()
-}
