+++ bzauth-rs/src/contracts/token.rs
@@ -1,0 +1,17 @@
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Default, Serialize, Deserialize)]
+#[serde(default)]
+pub struct Token {
+    pub access_token: Option<String>,
+    pub token_type: Option<String>,
+    pub refresh_token: Option<String>,
+    pub expires_in: Option<u64>,
+    pub scope: Option<String>,
+    pub id_token: Option<String>,
+
+    #[serde(flatten)]
+    pub others: HashMap<String, String>,
+}
