+++ bzauth-rs/src/tools/http/mod.rs
@@ -1,3 +1,4 @@
+pub mod cookie;
 pub mod request;
 pub mod response;
 pub mod status;
