+++ bzauth-rs/src/tools/mod.rs
@@ -1,19 +1,17 @@
 // Pub modules for the cogs
 
-pub mod cookie;
 pub mod error;
 pub mod http;
 pub mod routes;
 pub mod util;
 
-pub mod csrf;
-
 // Pub use statements for the modules
 
-pub use cookie::*;
-pub use csrf::*;
 pub use error::*;
+pub use http::cookie::*;
 pub use http::*;
 pub use routes::authorise::*;
 pub use routes::callback::*;
-pub use util::*;
+pub use routes::csrf::*;
+pub use util::try_async::*;
+pub use util::*; // Make the `TryFromAsync` trait available
