+++ bzauth-rs/src/tools/routes/csrf.rs
@@ -1,0 +1,5 @@
+use crate::tools::{CoreError, request::CoreRequest, response::CoreResponse};
+
+pub async fn csrf(_request: CoreRequest) -> Result<CoreResponse, CoreError> {
+    todo!("Need to implement CSRF token generation and validation");
+}
