+++ bzauth-rs/src/tools/routes/mod.rs
@@ -1,2 +1,3 @@
 pub mod authorise;
 pub mod callback;
+pub mod csrf;
